break out the providers router into a separate file
