+++ .changeset/six-sloths-listen.md
@@ -1,0 +1,5 @@
+---
+'@backstage/plugin-auth-backend': patch
+---
+
+Internal refactor to break out how the router is constructed
